extract message types to their own package
