+++ implementations/slims/orv/protocol/mt/messagetype.go
@@ -1,0 +1,77 @@
+package mt
+
+// MessageType enumerates the message types and the integers used to represent them in the Type field.
+type MessageType uint8
+
+const (
+	UNKNOWN MessageType = iota
+	Fault
+	Hello
+	HelloAck
+
+	Join
+	JoinAccept
+
+	Register
+	RegisterAccept
+
+	Merge
+	MergeAccept
+	Increment
+	IncrementAck
+
+	ServiceHeartbeat
+	ServiceHeartbeatAck
+
+	VKHeartbeat
+	VKHeartbeatAck
+
+	// client
+	Status
+	StatusResp
+	List
+	ListResp
+	Get
+	GetResp
+)
+
+// MessageTypeString returns the string representation of the given MessageType.
+// It is just a big switch statement.
+func (mt MessageType) String() string {
+	switch mt {
+	case Fault:
+		return "FAULT"
+	case Hello:
+		return "HELLO"
+	case HelloAck:
+		return "HELLO_ACK"
+	case Join:
+		return "JOIN"
+	case JoinAccept:
+		return "JOIN_ACCEPT"
+	case Register:
+		return "REGISTER"
+	case RegisterAccept:
+		return "REGISTER_ACCEPT"
+	case Merge:
+		return "MERGE"
+	case MergeAccept:
+		return "MERGE_ACCEPT"
+	case Increment:
+		return "INCREMENT"
+	case IncrementAck:
+		return "INCREMENT_ACK"
+	case ServiceHeartbeat:
+		return "SERVICE_HEARTBEAT"
+	case ServiceHeartbeatAck:
+		return "SERVICE_HEARTBEAT_ACK"
+	case VKHeartbeat:
+		return "VK_HEARTBEAT"
+	case VKHeartbeatAck:
+		return "VK_HEARTBEAT_ACK"
+	default:
+		return "UNKNOWN"
+	}
+}
+
+//#endregion MessageType
